Switch back to a lock-based WaitGroup implementation.
The existing implementation allowed state transitions to occur without
locking by keeping the state in an atomic. Unfortunately, this allowed
a pernicious edge case where the condition variable was notified while
a waiting thread held the lock, had observed an old state, and has yet
to call wait. This leads to deadlock, which is not desirable.

Holding the mutex both during state transitions and when waiting stops
the problem.